% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fdat}
\alias{fdat}
\title{This is an example data in NMRphasing}
\format{
A data frame with two columns, one is for NMR data in complex format, the other one is ppm
}
\usage{
fdat
}
\description{
This dataset contains sample data for NMRphasing.
}
\author{
Aixiang Jiang
}
\keyword{datasets}
