% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPC_DAOM.R
\name{MPC_DAOM}
\alias{MPC_DAOM}
\title{MPC_DAOM}
\usage{
MPC_DAOM(specdat)
}
\arguments{
\item{specdat}{A complex number vector of observed frequency domain data}
}
\value{
A matrix with phase and baseline corrected spectra, the 1st column is for absorption spectrum, while the 2nd column is for dispersion spectrum.
}
\description{
Multiple linear models with Minimization of difference between absolute area and ordinary area
}
\details{
This function is to process phase error correction through multiple linear models with Minimization of difference between absolute area and ordinary area,
followed by Polynomial baseline correction when necessary
}
\references{
Jiang A, Hanley JA, and Nadon R, 2021, 1D NMR Phase Error Correction with New Modeling Methods (in preparation)

Jiang A, Gravel A, Hanley JA, and Nadon R, 2021, Comparing phase error correction methods with NMR spike-in experiments (in preparation)

Liland KH, Almøy T, Mevik B (2010), Optimal Choice of Baseline
Correction for Multivariate Calibration of Spectra, Applied Spectroscopy 64, pp. 1007-1016.
}
\author{
Aixiang Jiang
}
\keyword{1D}
\keyword{NMR,}
\keyword{correction}
\keyword{phase}
