% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRphasing.R
\name{NMRphasing}
\alias{NMRphasing}
\title{NMRphasing}
\usage{
NMRphasing(
  specDatIn,
  absorptionOnly = FALSE,
  method = c("NLS", "MPC_DAOM", "MPC_EMP", "SPC_DAOM", "SPC_EMP", "SPC_AAM", "SPC_DSM")
)
}
\arguments{
\item{specDatIn}{Input spectrum data, which can be one of the four formats:
a vector of absorption specrtrum; a complex vector; a data matrix or a data frame with two columns of spectrum data,
which 1st column is for absorption spectrum, and 2nd column is for dispersion spectrum}

\item{absorptionOnly}{A logical variable to tell us if specDatIn is a a vector of absorption specrtrum, default is false}

\item{method}{One of phase correction method. There are seven methods right now, which are "NLS", "MPC_DAOM", "MPC_EMP", "SPC_DAOM", "SPC_EMP", "SPC_AAM", "SPC_DSM",
with NLS, non-linear shrinkage as default}
}
\value{
A matrix with phase and baseline corrected spectra, the 1st column is for absorption spectrum, while the 2nd column is for dispersion spectrum.
}
\description{
Phase error correction wrap up function
}
\details{
This is a wrap function to process phase error correction with seven different choices, followed by Polynomial baseline correction when necessary
}
\references{
de Brouwer, H. (2009). Evaluation of algorithms for automated phase correction of NMR spectra. J Magn Reson, 201, 230-238.

Dzakula, Z. (2000). Phase angle measurement from peak areas (PAMPAS). J Magn Reson, 146, 20-32.

Chen, L., Weng, Z., Goh, L., & Garland, M. (2002). An efficient algorithm for automatic phase correction of NMR spectra based on
entropy minimization. Journal of Magnetic Resonance, 158, 1-2.

Ernst, R. R. (1969). Numerical Hilbert transform and automatic phase correction in magnetic resonance spectroscopy.
Journal of Magnetic Resonance, 1, 7-26

Kristian Hovde Liland, Trygve Almøy, Bjørn-Helge Mevik (2010), Optimal Choice of Baseline
Correction for Multivariate Calibration of Spectra, Applied Spectroscopy 64, pp. 1007-1016.
}
\author{
Aixiang Jiang
}
\keyword{1D}
\keyword{NMR,}
\keyword{a}
\keyword{absolute}
\keyword{area}
\keyword{correction,}
\keyword{linear}
\keyword{minimization}
\keyword{model,}
\keyword{of}
\keyword{phase}
\keyword{single}
