% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPC_DAOM.R
\name{SPC_DAOM}
\alias{SPC_DAOM}
\title{SPC_DAOM}
\usage{
SPC_DAOM(specdat)
}
\arguments{
\item{specdat}{A complex number vector of observed frequency domain data}
}
\value{
A matrix with phase and baseline corrected spectra, the 1st column is for absorption spectrum, while the 2nd column is for dispersion spectrum.
}
\description{
A single linear model with Minimization of difference between absolute area and ordinary area
}
\details{
This function is to process phase error correction through a single linear model with Minimization of difference between absolute area and ordinary area,
followed by Polynomial baseline correction
}
\references{
This is our new algorithm

Kristian Hovde Liland, Trygve Almøy, Bjørn-Helge Mevik (2010), Optimal Choice of Baseline
Correction for Multivariate Calibration of Spectra, Applied Spectroscopy 64, pp. 1007-1016.
}
\author{
Aixiang Jiang
}
\keyword{1D}
\keyword{NMR,}
\keyword{correction}
\keyword{phase}
